changed to componenet
