+++ package.json
@@ -22,7 +22,7 @@
     "html2canvas": "^1.4.1",
     "jquery": "^3.6.1",
     "jspdf": "^2.5.1",
-    "mongoose": "^6.5.2",
+    "mongoose": "^6.5.4",
     "next": "12.2.4",
     "next-auth": "^4.10.3",
     "next-compose-plugins": "^2.2.1",
